move implementaiton of persist from DataSet.scala back into SmvDataSet.scala (#381)
